Use MT_Vector4 for horizon and zenith colors.
Using vector of 4d allow to use the alpha in the bacnground render.
This case happen for the ImageRender horizon color which was always
a 4d contrary to the horizon color exposed in KX_WorldInfo.

So the issue was that the alpha value managed into ImageRender was
unused in KX_WorldInfo.

To fix this issue KX_WorldInfo fully switch to 4d internally and
externally (python API) for the horizon and zenith colors.
Unfortunatly in the python API these value became 4d mathutils vector
instead of mathutils color.

Fix bug #431.